+++ ai-logic/firebase-ai/src/main/kotlin/com/google/firebase/ai/Chat.kt
@@ -87,7 +87,7 @@ public class Chat(
           break
         }
         val functionResponsePart = functionCallParts.map { model.executeFunction(it) }
-        tempHistory.add(Content("function", functionResponsePart))
+        tempHistory.add(Content("user", functionResponsePart))
       }
       history.addAll(tempHistory)
       return response
@@ -220,8 +220,7 @@ public class Chat(
         if (model.requestOptions.autoFunctionCallingTurnLimit < ++turns) {
           throw RequestTimeoutException("Request took too many turns", history = tempHistory)
         }
-        val functionResponses =
-          Content("function", functionCallParts.map { model.executeFunction(it) })
+        val functionResponses = Content("user", functionCallParts.map { model.executeFunction(it) })
         tempHistory.add(Content("model", functionCallParts))
         tempHistory.add(functionResponses)
         model
@@ -238,8 +237,8 @@ public class Chat(
   }
 
   private fun Content.assertComesFromUser() {
-    if (role !in listOf("user", "function")) {
-      throw InvalidStateException("Chat prompts should come from the 'user' or 'function' role.")
+    if (role != "user") {
+      throw InvalidStateException("Chat prompts should come from the 'user' role.")
     }
   }